Move module in correct path
