+++ rpcs3/Emu/Cell/Modules/cellCrossController.cpp
