+++ rpcs3/Emu/Cell/Modules/cellSysutilNpEula.cpp
